Fix: Error de encoding tablas importadas

+++ app/Jobs/ProcessImports.php
@@ -58,10 +58,13 @@ class ProcessImports implements ShouldQueue
         DB::table($this->table)->truncate();
 
         // Cargo el archivo
-        $sql = "LOAD DATA LOCAL INFILE '" . addslashes($prefix . $this->file) . "' INTO TABLE $this->table FIELDS TERMINATED BY ';' IGNORE 1 LINES";
+        $sql = "LOAD DATA LOCAL INFILE '" . addslashes($prefix . $this->file) . "' INTO TABLE $this->table CHARACTER SET LATIN1 FIELDS TERMINATED BY ';' IGNORE 1 LINES";
         DB::connection()->getpdo()->exec($sql); // https://stackoverflow.com/a/44426882
 
-        // Elimino el archivo
-        Storage::disk('local')->delete($this->file);
+        // Si no estamos trabajando en local
+        if (!\App::environment('local')) {
+            // Elimino el archivo
+            Storage::disk('local')->delete($this->file);
+        }
     }
 }
